LEPL-23: промежуточные изменения по таске. Перенёс маппер AdminCourseMapper.
